Create custom packages for the project
There are multiple classes that are not categorized into packages.

Creating custom packages improves the standardization of the code

As a step towards making the code better, let's extract those classes that are similar and package them into the same package
