+++ src/main/java/OUTPUT.txt
@@ -1,0 +1,1 @@
+[E] [] fuck owens mum (from : Oct 2 3409 to: Oct 23 4390)
