+++ src/main/java/Tasks/Deadlines.java
@@ -1,6 +1,15 @@
-public class ToDos extends Task {
-    ToDos(String name, boolean isDone) {
+package Tasks;
+
+import java.time.LocalDate;
+import java.time.format.DateTimeFormatter;
+import java.time.format.DateTimeParseException;
+
+public class Deadlines extends Task {
+    private LocalDate by;
+
+    public Deadlines(String name, String by, boolean isDone) throws DateTimeParseException {
         super(name, isDone);
+        this.by = LocalDate.parse(by.trim());
     }
 
     /*
@@ -9,13 +18,13 @@ public class ToDos extends Task {
      */
     @Override
     public String toString() {
-        return "[T] " + super.toString();
+        return "[D] " + super.toString() + "(by : " + this.by.format(DateTimeFormatter.ofPattern("MMM d yyyy")) + ")";
     }
 
     /*
-    Method name: markDone
-    Description: Checks the specified task
-     */
+Method name: markDone
+Description: Checks the specified task
+ */
     @Override
     public void markDone() {
         super.markDone();
